fix: Restructure for clean include and src
