+++ CMakeLists.txt
@@ -26,7 +26,7 @@ target_link_libraries(gtfs-rt protobuf::libprotobuf)
 # these instructions search the directory tree when CMake is
 # invoked and put all files that match the pattern in the variables
 # `sources` and `data`
-file(GLOB_RECURSE sources      exe/backend/*.cc)
+file(GLOB_RECURSE sources exe/backend/src/*.cc)
 # you can use set(sources src/main.cpp) etc if you don't want to
 # use globbing to find files automatically
 
@@ -36,10 +36,10 @@ file(GLOB_RECURSE sources      exe/backend/*.cc)
 
 # add the data to the target, so it becomes visible in some IDE
 
-file(GLOB_RECURSE tup-backend-src exe/backend/*.cc src/*.cc include/tup/*.h)
+file(GLOB_RECURSE tup-backend-src exe/backend/src/*.cc exe/backend/include/*.h)
 add_executable(tup-backend ${tup-backend-src})
 target_link_libraries(tup-backend boost-json conf boost fmt utl geo web-server nigiri tbb curl gtfs-rt)
-target_include_directories(tup-backend PRIVATE exe/backend/include include/tup)
+target_include_directories(tup-backend PRIVATE exe/backend/include)
 
 # include custom configuration from motis-project
 include(cmake/buildcache.cmake)
