+++ exe/backend/include/feed_updater.h
@@ -32,6 +32,10 @@ private:
   std::thread worker_;
   std::atomic<bool> running_{true};
   PredictionMethod predictionMethod_;
+  static size_t WriteCallback(void* contents, size_t size, size_t nmemb, void* userp) {
+    ((std::string*)userp)->append((char*)contents, size * nmemb);
+    return size * nmemb;
+  }
 };
