+++ exe/backend/src/feed_updater.cc
@@ -4,11 +4,6 @@
 #include <chrono>
 #include <thread>
 
-size_t FeedWriteCallback(void* contents, size_t size, size_t nmemb, void* userp) {
-  ((std::string*)userp)->append((char*)contents, size * nmemb);
-  return size * nmemb;
-}
-
 FeedUpdater::~FeedUpdater() {
   stop();
 }
@@ -43,7 +38,7 @@ bool FeedUpdater::downloadFeed() {
   if (!curl) return false;
 
   curl_easy_setopt(curl, CURLOPT_URL, url_.c_str());
-  curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, FeedWriteCallback);
+  curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, WriteCallback);
   curl_easy_setopt(curl, CURLOPT_WRITEDATA, &protobuf_data);
   CURLcode res = curl_easy_perform(curl);
   curl_easy_cleanup(curl);
